Backend/requirements.txt
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.8.2
bs4==0.0.1
cachelib==0.9.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.5
cssselect==1.2.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.18.0
filelock==3.12.4
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.42.1
fqdn==1.5.1
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==7.0.0
importlib-resources==6.0.1
incremental==22.10.0
ipykernel==6.27.1
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.5.0
lxml==4.9.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mysql-connector-python==9.0.0
nest-asyncio==1.5.8
outcome==1.3.0.post0
packaging==23.2
parsel==1.8.1
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.41
Protego==0.3.0
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.17.2
pymongo==4.6.1
pyOpenSSL==23.3.0
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1

pyzmq==25.1.2
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
selenium==4.15.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
tldextract==5.0.1
tornado==6.4
traitlets==5.14.0
trio==0.23.1
trio-websocket==0.11.1
Twisted==22.10.0

urllib3==2.0.7
w3lib==2.1.2
wcwidth==0.2.12
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
zipp==3.17.0
zope.interface==6.1
